Move Fake DNS to a separate package
