+++ app/dns/fakedns/fakedns.go
@@ -1,0 +1,3 @@
+package fakedns
+
+//go:generate go run v2ray.com/core/common/errors/errorgen
