+++ features/dns/fakedns.go
@@ -2,7 +2,7 @@ package dns
 
 import "v2ray.com/core/common/net"
 
-type FakeDns interface {
+type FakeDnsEngine interface {
 	GetFakeIPForDomain(domain string) []net.Address
 	GetDomainFromFakeDNS(ip net.Address) string
 }
